requirements.txt
asgiref==3.8.1
Django==5.1.3
django-filter==24.3
django-galleryfield==2.2.0
pillow==11.0.0
sorl-thumbnail==12.11.0
sqlparse==0.5.2
